reworked points form

+++ frontend/src/components/ProPoints/pointsForm.js
@@ -7,33 +7,37 @@ import Typography from "@mui/material/Typography";
 
 import axios from "axios";
 import qs from "qs";
+import { Controller, useForm } from "react-hook-form";
 
 export default function AddPoints() {
   const [msg, setMsg] = React.useState("");
-  const [event, setEvent] = React.useState("");
   const [disable, setDisable] = React.useState(false);
+  const { control, handleSubmit, watch, reset, resetField, setError, clearErrors, formState : {errors} } = useForm({
+    defaultValues: {
+      points: 1,
+      courseId: '',
+      description: '',
+      EventId: ''
+    },
+  });
+  const watchEventId = watch("EventId");
 
-  const handleSubmit = (eve) => {
-    eve.preventDefault();
+  const onSubmit = (data) => {
     setDisable(true);
-    let data = new FormData(eve.currentTarget);
-    data = {
-      points: data.get("propoint"),
-      courseId: data.get("courseId"),
-      description: data.get("description") || event,
-      EventId: data.get("eventId"),
-    };
-
     axios
-      .post("/api/propoint", qs.stringify(data))
+      .post("/api/propoint", qs.stringify(data), {timeout: 5000})
       .then(function (res) {
         setDisable(false);
-        eve.target.reset();
         setMsg(
           `Successfully added ProPoint : ${res.data.id} - ${res.data.description}`
         );
       })
       .catch((err) => {
+        if(err.request){
+          setMsg("Unable to establish remote server connection.");
+          setDisable(false);
+          return;
+        }
         // Unauthorized
         if (err.response.status === 401) {
           localStorage.clear();
@@ -43,35 +47,56 @@ export default function AddPoints() {
           setMsg("Please enter in the correct Event/Course ID!");
           setDisable(false);
         }
-      });
+    });
+
+    reset({
+      EventId: '',
+      courseId: '',
+      description: '',
+      points: 1
+    });
   };
 
-  const handleEvent = (event) => {
+  React.useEffect( () => {
+    if(!watchEventId){
+      return ;
+    }
     axios
-      .get("/api/event", {
-        params: {
-          id: event.target.value,
-        },
-      })
-      .then(function (res) {
-        if (res.data) {
-          setEvent(res.data.event);
-        } else {
-          setEvent("No Matching event ID");
-        }
-      })
-      .catch(function (err) {
-        if (err.response.status === 401) {
-          localStorage.clear();
-          setMsg("Unauthorized. Please refresh the page and login!");
-        }
-      });
-  };
+    .get("/api/event", {
+      params: {
+        id: watchEventId,
+      },
+      timeout: 5000, // 5 seconds
+    })
+    .then(function (res) {
+      if (res.data) {
+        resetField('description', {keepError: false, defaultValue: res.data.event});
+        clearErrors('EventId');
+      } else {
+        resetField('description', {defaultValue: "No Matching event ID"});
+        setError('EventId', {
+          type: 'manual',
+          message: 'This is not a registered Event!'
+        });
+      }
+    })
+    .catch(function (err) {
+      if(err.request){
+        setMsg('Unable to establish database connection');
+        return;
+      }
+      if (err.response.status === 401) {
+        localStorage.clear();
+        setMsg("Unauthorized. Please refresh the page and login!");
+      }
+    });
+  }, [watchEventId])
+
 
   return (
     <Box
       component="form"
-      onSubmit={handleSubmit}
+      onSubmit={handleSubmit(onSubmit)}
       noValidate
       sx={{
         margin: 2,
@@ -90,57 +115,76 @@ export default function AddPoints() {
         alignItems="center"
       >
         <Grid item xs = {12} md={12}>
-          <TextField
-            margin="normal"
+          <Controller
+          name = "EventId"
+          control = {control}
+          rules= {{ 
+            required: true,
+            pattern: /^\d+$/,
+            message: "Please enter a registed Event ID!"
+          }}
+          render = { ({field, fieldState}) => 
+            <TextField
+            {...field}
             type="number"
-            required
-            fullWidth
-            onChange={handleEvent}
-            error={Boolean(event == "No Matching event ID")}
-            id="id"
             label="Event ID"
-            name="eventId"
-            autoFocus={true}
-          />
+            error= {fieldState.error}
+            helperText={errors?.EventId?.message}
+            fullWidth
+            required
+            />
+        }/>
         </Grid>
         <Grid item xs={12}>
-          <TextField
-            margin="normal"
-            type="number"
-            required
-            fullWidth
-            id="courseId"
-            label="Course ID"
-            name="courseId"
+          <Controller
+          name = "courseId"
+          control={control}
+          rules = {{
+            required : true,
+            minLength: 4,
+            pattern: /^\d+$/,
+            message: "I.e. Robotics Lab is ECE '3331'"
+          }}
+          render = { ({field, fieldState}) => 
+            <TextField
+            {...field}
+            type= "number"
+            label= "Course Number"  
             placeholder="3331"
-            autoFocus
-          />
-        </Grid>
-        <Grid item xs={12}>
-          <TextField
-            margin="normal"
+            error= {fieldState.error}
+            helperText= "I.e. Robotics Lab is ECE '3331'"
             fullWidth
-            multiline
-            id="description"
-            label="Event Title or Description"
-            name="description"
-            placeholder={event}
-            error={Boolean(event == "No Matching event ID")}
-            autoFocus
-          />
-        </Grid>
-        <Grid item xs={12}>
-          <TextField
-            margin="normal"
-            type="number"
             required
-            fullWidth
-            id="points"
-            label="ProPoints"
-            name="propoint"
-            defaultValue="1"
+            />
+          }/>
+        </Grid>
+        <Grid item xs={12}>
+          <Controller
+          name= "description"
+          control = {control}
+          render= { ({field, fieldState}) => 
+            <TextField
+            {...field}
+            label= "Event Title or Description"
             autoFocus
-          />
+            multiline
+            />
+          }/>
+        </Grid>
+        <Grid item xs={12}>
+          <Controller
+          name="points"
+          control= {control}
+          render= { ({field}) => 
+            <TextField
+            {...field}
+            type="number"
+            label="ProPoints"
+            fullWidth
+            defaultValue={field.defaultValue}
+            required
+            />
+          }/>
         </Grid>
         <Grid item xs={12} sm={6}>
           <Typography variant="body1">{msg}</Typography>
